linux/requirements.txt
argparse
google-generativeai
colorama